Deleted old log class and log table model. Cleaned up unused methods, variables and includes. Replaced the QString with network datagram in iodeviceform .

+++ incl/domain/log.h
@@ -6,7 +6,7 @@
 struct BsfLog {
     int id;
     int logType;
-    int logDateTime;
+    qint64 logDateTime;
     QString log;
 };
